organized_data_management_api/requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.25.2
requests>=2.31.0

# Database and ORM
sqlalchemy>=2.0.23
asyncpg>=0.29.0
aiomysql>=0.2.0
psycopg2-binary>=2.9.0

# AI and ML dependencies
openai>=1.3.7
litellm>=1.17.0
tiktoken>=0.5.2

# AI Agents package (used extensively in the code)
openai-agents>=0.3.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0

# Optional: For enhanced logging
structlog>=23.2.0

# Optional: For better error handling
sentry-sdk[fastapi]>=1.38.0
